Define local plugins in $HOME/.vimrc.plugins
The plugin code in vimrc was a bit ugly so I refactored it. After doing
so, we got an immediate new feature: ability to load plugins from any
(well known) vim source. In this case $HOME/.vimrc.plugins. This works
for testing plugins locally and not have to meddle with vimrc and the
git repository.